Changing location of export splats function for smooth run of unit tests
